Updated deflection logic (possibly bugged, disabled for apple, as it didn't work)
The logic for projectiles deflecting off the ground has been updated to not just take the normal direction, but to take the velocity direction reflected on the normal.
It looks like it works, but it may be broken. It didn't look like it worked for the apple, so i disabled it for that.