Updated player physics
Changed jump logic to use impulse force instead of default force, which is force force (force over time instead of instantanious). With this change, the character seems to no longer accelerate over the duration of the jump, but instead keeps a constant speed as long as the button is held.

I also updated the different movement variables in the prefab, and added a seperate friction variable for the air