Added Fairy enemy
Added a fairy enemy which flies around.
If the player comes close to the fairy, it gets angry and goes in range to shoot. It keeps its distance and tries to stay at a diagonal angle to the player.
If the player gets far enough away, the fairy returns home.

It's also supposed to shoot projectiles, but i haven't implemented that yet.