Renamed bulletSpin to bullet
Renamed the bulletSpin script to just bullet.
Earlier in development, bullet was supposed to spin, and was thus called bulletSpin, but I scrapped it, because the idea I had to make it spin turned out not to work, and theres no reason to overcomplicate it :/
While I had renamed the script in the code to bullet already, I apparently hadn't changed the filename too. This commit fixes that.