Fairy: burstProjectileDelay is no longer included in burstTimeMin/MaxBase
When the Fairy enemy shoots a burst of projectiles, it waits burstProjectileDelay between each shot, but it also waited burstProjectileDelay after the burst, but that is no longer the case.

The base delay for each of the Fairy enemy prefabs has been increased to compensate, so their behaviour should still be the same.