Made the Fairys movement vector stay short of it's too short
The Fairys movement vector is made up of a few combined behaviour vectors. This combined vector was then normalized and multiplied with speed to find make the final movement vector, however, when these vectors almost cancelled out to zero, the movement could become janky, as small changes in the vectors could make a bit difference.
To try to combat this, I added a new variable, lowSpeedThreshold, where, if the combined vectors are shorter than this length, they will be scaled to be shorter than a regular normalization, so the Fairy will end up slowing down, instead of rapidly changing directions.