Updated projectiles to make sure they don't collide with their team
Updated projectiles so that they don't hit entities on their own team, as in, an enemy projectile wont hit other enemies, and an ally projectile wont hit allies.

This is done by using two layers, an "Ally Entity" layer and an "Enemy Entity" layer, and projectiles are then told whether they are allied or enemies, and exclude the relevant layers from their collision detection.

To do this, I made quite a lot of spaghetti, and it became kindof a mess, and I also deleted the "Collidable" layer, as I believe it was unused? I'm sorry for the mess.