Fixed being able to jump off of projectiles
Fixed a bug where the player could sometimes jump on projectiles.
This was done in a kind of janky way by adding a tag to all projectiles, marking them as "NotGround", and checking if the ground the player/entity landed on is tagged with that tag before marking it grounded.