Fixed deflection logic in gingerbreadMan, and added it to apple
The old deflection logic used for gingerbreadMan on hit seemed to work, but it didn't work for apple, which was a bad sign.
Now I made a fixed, working version for the apple, and implemented it in the gingerbreadMan too. This version also uses the relative velocity between the colliders, leading to more realistic results when they strike a moving target.