Added Fairy Rat Enemy
The Rat Fairy has arrived!
This special variation of the Fairy Enemy possesses a unique Cheese projectile, while behaves, and currently looks, like an apple. The special Rat Fairy also never truly gets defeated, as it has a special death animation, which, instead of dying, has it surrender into the sky, when it loses all of its health.